+++ app.h
@@ -1,0 +1,5 @@
+char *handleRequest(char *raw_request);
+void indexPage(Response *resp, Request *req);
+void helloPage(Response *resp);
+void dinosaurIndexPage(Response *resp, Request *req);
+void dinosaurShowPage (Response *resp, Request *req);
